OCR-Server/requirements.txt
Flask
Werkzeug
gunicorn
pytesseract
Pillow
pdf2image
opencv-python
numpy
